Fix tests when no controller repo is set
This also extracts the helper to do so into a separate function to avoid
duplicating this logic.
