+++ extensions/ql-vscode/test/vscode-tests/minimal-workspace/databases/db-panel-rendering.test.ts
@@ -13,6 +13,7 @@ import { DbTreeViewItem } from "../../../../src/databases/ui/db-tree-view-item";
 import { ExtensionApp } from "../../../../src/common/vscode/vscode-app";
 import { createMockExtensionContext } from "../../../factories/extension-context";
 import { createDbConfig } from "../../../factories/db-config-factories";
+import { mockConfiguration } from "../../utils/configuration-helpers";
 
 describe("db panel rendering nodes", () => {
   const workspaceStoragePath = join(__dirname, "test-workspace-storage");
@@ -42,6 +43,14 @@ describe("db panel rendering nodes", () => {
 
   beforeEach(async () => {
     await ensureDir(workspaceStoragePath);
+
+    mockConfiguration({
+      values: {
+        "codeQL.variantAnalysis": {
+          controllerRepo: "github/codeql",
+        },
+      },
+    });
   });
 
   afterEach(async () => {
